Now able to translate between written word
Just so you know, this was written under 2 hours by a noob named: me!
Damn i'm proud of myself.
Anyway. The TranslateActivity passes a translation to each translate
TranslateHolder. The TranslateHolder splits the translation into a word
and a translation and passes them to the two translateFragments.

The app is now useful, the day before school start. Just in time.
Phew...
/JohnTheDaniel